+++ modules/liquidity-pools/synthetic/src/mock.rs
@@ -16,7 +16,7 @@ use primitives::{Balance, CurrencyId, LiquidityPoolId};
 use traits::BaseLiquidityPoolManager;
 
 pub type BlockNumber = u64;
-pub type AccountId = u32;
+pub type AccountId = u64;
 
 ord_parameter_types! {
 	pub const One: AccountId = 0;
